[LinkFlight]: Version 2 & 3
[Change]
1. add the angle of the planes' pictures.
2. add the airports' location as a JSON file(It will be transferred to the online JSON file in the future).
3. add the details of the airplane.

[Symptom]

[Cause]

[Solution]

[Issue type]
develop

[others]
We use two flight API providers to make this application. The responsibility of OpenSkyNetwork API is to check the location of the airplanes once the user drag, zooms In and Out, retrieves the data, and shows marks on the map. FlightAware API for fetching the details of the airplane by callsign.